Move some code to utils file
