+++ graphql_compiler/schema_transformation/utils.py
@@ -2,13 +2,14 @@
 import string
 
 from graphql import build_ast_schema
-from graphql.language.ast import NamedType
+from graphql.language import ast as ast_types
 from graphql.language.visitor import Visitor, visit
 from graphql.type.definition import GraphQLScalarType
 from graphql.utils.assert_valid_name import COMPILED_NAME_PATTERN
+from graphql.validation import validate
 import six
 
-from ..exceptions import GraphQLError
+from ..exceptions import GraphQLError, GraphQLValidationError
 
 
 class SchemaTransformError(GraphQLError):
@@ -238,7 +239,7 @@ class CheckQueryTypeFieldsNameMatchVisitor(Visitor):
             field_name = node.name.value
             type_node = node.type
             # NamedType node may be wrapped in several layers of NonNullType or ListType
-            while not isinstance(type_node, NamedType):
+            while not isinstance(type_node, ast_types.NamedType):
                 type_node = type_node.type
             queried_type_name = type_node.name.value
             if field_name != queried_type_name:
@@ -283,3 +284,84 @@ def check_ast_schema_is_valid(ast):
 
     query_type = get_query_type_name(schema)
     visit(ast, CheckQueryTypeFieldsNameMatchVisitor(query_type))
+
+
+def is_property_field(selection):
+    """Return True if selection is a property field, False if a vertex field or type coercion.
+
+    Args:
+        selection: type Field or InlineFragment. An element occuring inside a SelectionSet
+
+    Returns:
+        True if the selection is a property field, False if it's a vertex field or type coercion
+    """
+    if isinstance(selection, ast_types.InlineFragment):
+        return False
+    if isinstance(selection, ast_types.Field):
+        if (
+            selection.selection_set is None or
+            selection.selection_set.selections is None or
+            selection.selection_set.selections == []
+        ):
+            return True
+        else:
+            return False
+    else:
+        raise AssertionError(
+            u'Input selection "{}" is not of type Field or InlineFragment.'.format(selection)
+        )
+
+
+class CheckQueryIsValidToSplitVisitor(Visitor):
+    """Check the query only has supported directives, and its fields are correctly ordered."""
+    # This is very restrictive for now. Other cases (e.g. tags not crossing boundaries) are
+    # also ok, but temporarily not allowed
+    supported_directives = frozenset(('filter', 'output', 'optional'))
+
+    def enter_Directive(self, node, *args):
+        """Check that the directive is supported."""
+        if node.name.value not in self.supported_directives:
+            raise GraphQLValidationError(
+                u'Directive "{}" is not yet supported, only "{}" are currently '
+                u'supported.'.format(node.name.value, self.supported_directives)
+            )
+
+    def enter_SelectionSet(self, node, *args):
+        """Check property fields occur before vertex fields and type coercions in selection."""
+        past_property_fields = False  # Whether we're seen a vertex field
+        for field in node.selections:
+            if is_property_field(field):
+                if past_property_fields:
+                    raise GraphQLValidationError(
+                        u'In the selections {}, the property field {} occurs after a vertex '
+                        u'field or a type coercion statement, which is not allowed, as all '
+                        u'property fields must appear before all vertex fields.'.format(
+                            node.selections, field
+                        )
+                    )
+            else:
+                past_property_fields = True
+
+
+def check_query_is_valid_to_split(schema, query_ast):
+    """Check the query is valid for splitting.
+    In particular, ensure that the query validates against the schema, does not contain
+    unsupported directives, and that in each selection, all property fields occur before all
+    vertex fields.
+    Args:
+        schema: GraphQLSchema object
+        query_ast: Document
+    Raises:
+        GraphQLValidationError if the query doesn't validate against the schema, contains
+        unsupported directives, or some property field occurs after a vertex field in some
+        selection
+    """
+    # Check builtin errors
+    built_in_validation_errors = validate(schema, query_ast)
+    if len(built_in_validation_errors) > 0:
+        raise GraphQLValidationError(
+            u'AST does not validate: {}'.format(built_in_validation_errors)
+        )
+    # Check no bad directives and fields are in order
+    visitor = CheckQueryIsValidToSplitVisitor()
+    visit(query_ast, visitor)
